backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
psycopg2-binary>=2.9,<3.0
python-decouple>=3.8,<4.0
dj-database-url>=2.1,<3.0
Pillow>=10.0,<11.0
google-auth>=2.23,<3.0
google-auth-oauthlib>=1.1,<2.0
google-auth-httplib2>=0.1,<1.0
